moved new_solver tests to solver
